refactor: move getParentsAlias FROM UserService TO UserUtils
